Interface refactor.
Was not familier with Java interface conventions. See http://stackoverflow.com/questions/2814805/java-interfaces-implementation-naming-convention .
