--- memo-list-magic-core/src/main/java/com/redhat/mlm/IEmailStoreFactory.java
@@ -1,7 +1,0 @@
-package com.redhat.mlm;
-
-import javax.mail.Store;
-
-public interface IEmailStoreFactory {
-	Store getEmailStore();
-}
