+++ memo-list-magic-core/src/test/java/com/redhat/mlm/EmailStoreFactoryTest.java
@@ -6,22 +6,22 @@ import static org.junit.Assert.*;
 public class EmailStoreFactoryTest {
 	@Test
 	public void testFactoryReturnsNonNullStore(){
-		EmailStoreFactory factory = new EmailStoreFactory("host", "user", "pass");
+		IMAPEmailStoreFactory factory = new IMAPEmailStoreFactory("host", "user", "pass");
 		assertNotNull(factory.getEmailStore());
 	}
 	
 	@Test(expected = NullPointerException.class)
 	public void testFactoryThrowsErrorWhenCreatedWithNullHost(){
-		new EmailStoreFactory(null, "user", "pass");
+		new IMAPEmailStoreFactory(null, "user", "pass");
 	}
 	
 	@Test(expected = NullPointerException.class)
 	public void testFactoryThrowsErrorWhenCreatedWithNullUser(){
-		new EmailStoreFactory("host", null, "pass");
+		new IMAPEmailStoreFactory("host", null, "pass");
 	}
 	
 	@Test(expected = NullPointerException.class)
 	public void testFactoryThrowsErrorWhenCreatedWithNullPassword(){
-		new EmailStoreFactory("host", "user", null);
+		new IMAPEmailStoreFactory("host", "user", null);
 	}
 }
